# Conflicts:
#	TheElement/app/src/main/assets/db/Element.script
#	TheElement/app/src/main/java/fifthelement/theelement/business/services/SongService.java
#	TheElement/app/src/main/java/fifthelement/theelement/persistence/hsqldb/AlbumPersistenceHSQLDB.java
#	TheElement/app/src/main/java/fifthelement/theelement/persistence/hsqldb/AuthorPersistenceHSQLDB.java
#	TheElement/app/src/main/java/fifthelement/theelement/persistence/hsqldb/SongPersistenceHSQLDB.java
#	TheElement/app/src/main/java/fifthelement/theelement/presentation/fragments/SettingFragment.java
#	TheElement/app/src/main/java/fifthelement/theelement/presentation/services/DrawerService.java
#	TheElement/app/src/main/java/fifthelement/theelement/presentation/services/MusicService.java
#	TheElement/app/src/test/java/fifthelement/theelement/business/SongServiceTest.java

+++ TheElement/app/src/main/java/fifthelement/theelement/persistence/hsqldb/AlbumPersistenceHSQLDB.java
@@ -11,7 +11,6 @@ import java.util.List;
 import java.util.UUID;
 
 import fifthelement.theelement.application.Persistence;
-import fifthelement.theelement.application.Services;
 import fifthelement.theelement.objects.Album;
 import fifthelement.theelement.objects.Author;
 import fifthelement.theelement.objects.Song;
@@ -35,7 +34,7 @@ public class AlbumPersistenceHSQLDB implements AlbumPersistence {
         final String authorUUID = rs.getString("authorUUID");
         Author author = null;
         if(authorUUID != null)
-            author = Persistence.getAuthorPersistence().getAuthorByUUID(UUID.fromString(authorUUID));
+            author = new Author(UUID.fromString(authorUUID), "");
         final List<Song> songs = null;
         final int numPlayed = rs.getInt("numPlayed");
         return new Album(albumUUID, albumName, author, songs, numPlayed);
@@ -76,9 +75,11 @@ public class AlbumPersistenceHSQLDB implements AlbumPersistence {
             String uuidString = uuid.toString();
             st.setString(1, uuidString);
 
+            Album album = null;
             final ResultSet rs = st.executeQuery();
-            rs.next();
-            final Album album = fromResultSet(rs);
+            if(rs.next()) {
+                album = fromResultSet(rs);
+            }
             rs.close();
             st.close();
 
@@ -118,13 +119,17 @@ public class AlbumPersistenceHSQLDB implements AlbumPersistence {
         if(album == null)
             throw new IllegalArgumentException("Cannot update a null album");
         try {
-            final PreparedStatement st = c.prepareStatement("UPDATE albums SET albumName = ?, numPlayed = ? WHERE albumUUID = ? ");
-            st.setString(1, album.getName());
-            st.setInt(2, album.getNumPlayed());
-            st.setString(3, album.getUUID().toString());
+            if(albumExists(album)) {
+                final PreparedStatement st = c.prepareStatement("UPDATE albums SET albumName = ?, numPlayed = ? WHERE albumUUID = ?");
+                st.setString(1, album.getName());
+                st.setInt(2, album.getNumPlayed());
+                st.setString(3, album.getUUID().toString());
 
-            st.executeUpdate();
-            return true;
+                st.executeUpdate();
+
+                return true;
+            }
+            return false;
         } catch (final SQLException e) {
             throw new PersistenceException(e);
         }
@@ -143,16 +148,18 @@ public class AlbumPersistenceHSQLDB implements AlbumPersistence {
         if(uuid == null)
             throw new IllegalArgumentException("Cannot delete album with a null UUID");
         try {
-            final PreparedStatement st = c.prepareStatement("DELETE FROM albums WHERE albumUUID = ?");
-            st.setString(1, uuid.toString());
+            if(getAlbumByUUID(uuid) != null) {
+                final PreparedStatement st = c.prepareStatement("DELETE FROM albums WHERE albumUUID = ?");
+                st.setString(1, uuid.toString());
 
-            st.executeUpdate();
-
+                st.executeUpdate();
+                removed = true;
+            }
         } catch (final SQLException e) {
             throw new PersistenceException(e);
         }
 
-        return  removed;
+        return removed;
     }
 
     @Override
